% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{append.likelihoods}
\alias{append.likelihoods}
\title{Append (normalized) likelihoods to a data.frame containing log-likelihoods.}
\usage{
append.likelihoods(d)
}
\arguments{
\item{d}{A data.frame containing a column of log-likelihoods named loglike.}
}
\value{
The augmented data.frame.
}
\description{
Append (normalized) likelihoods to a data.frame containing log-likelihoods.
}

