% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cols2vmat}
\alias{cols2vmat}
\title{Transform a data.frame carrying grid-oriented data into a list of grid
vectors and a data matrix.}
\usage{
cols2vmat(d)
}
\arguments{
\item{d}{A data.frame where the first \code{n} columns are vectors of grid
coordinates, and the final 2 columns contain log-likelihoods and normalized
likelihoods. The grid coorindate columns should be in order from least
rapidly varying to most rapidly varying.}
}
\value{
a list of length n; the first n-1 entries are the grid values in each
  direction, and the last entry is the (n-1) dimensional array of the data
  values at each grid point.
}
\description{
Transform a data.frame carrying grid-oriented data into a list of grid
vectors and a data matrix.
}

