% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_input.R
\name{cosmo.scan}
\alias{cosmo.scan}
\title{Read data from a CosmoSIS format linearized matrix sample file,
e.g. a power spectrum file.}
\usage{
cosmo.scan(dirname, filename, quiet = TRUE)
}
\arguments{
\item{dirname}{The name of the directory from which we read.}

\item{filename}{The name of the file we will read.}

\item{quiet}{Passed to \code{base::scan}.}
}
\value{
The vector of values.
}
\description{
Read data from a CosmoSIS format linearized matrix sample file,
e.g. a power spectrum file.
}

