% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cosmosis_tbl.R
\name{cosmosis_tbl-class}
\alias{cosmosis_tbl-class}
\title{`cosmosis_tbl` class}
\description{
The `cosmosis_tbl` class is a subclass of [`tbl_df`][tibble::tbl_df()],
created in order to have different default behaviour.

`cosmosis_tbl` provides functions for a few generic methods:
    summary: provides point estimates and credible regions for each
             sampled parameter
}
\section{Properties of `cosmosis_tbl`}{


Objects of class `tbl_df` have:
* A `class` attribute of `c("cosmosis_tbl", tbl_df", "tbl", "data.frame")`.
}

