% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{emcee.convergence.plot}
\alias{emcee.convergence.plot}
\title{Visualize the convergence of Emcee sampler output}
\usage{
emcee.convergence.plot(
  x,
  panels = TRUE,
  numbers = FALSE,
  nmin = 1,
  nmax = max(x$sample),
  walkers = 1:max(x$walker)
)
}
\arguments{
\item{x}{: a dataframe made from emcee sampler output}

\item{panels}{: logical, if TRUE use panels to separate walkers}

\item{numbers}{: logical, if TRUE print sample numbers on points}

\item{nmin}{: integer, first sample to plot}

\item{nmax}{: integer, last sample to plot}

\item{walkers}{: integer vector, indices of walkers to plot. If NULL, plot
all walkers}
}
\value{
a ggplot object
}
\description{
Visualize the convergence of Emcee sampler output
}
