% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{emcee.count.walkers}
\alias{emcee.count.walkers}
\title{emcee.count.walkers Return the number of walkers used for this EMCEE run.}
\usage{
emcee.count.walkers(txt)
}
\arguments{
\item{txt}{Starting lines from the EMCEE output file}
}
\value{
The number of walkers
}
\description{
emcee.count.walkers Return the number of walkers used for this EMCEE run.
}
