% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{emcee.read}
\alias{emcee.read}
\title{emcee.read}
\usage{
emcee.read(fname)
}
\arguments{
\item{fname}{The name of the CosmoSIS MCMC sampler output file to be read.
making the data frame}
}
\value{
a CosmoSIS MCMC dataframe
}
\description{
emcee.read
}
