% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find.contours}
\alias{find.contours}
\title{Find values at which given KDE takes specified values}
\usage{
find.contours(kde, levels = c(0.6826895, 0.9544997, 0.9973002))
}
\arguments{
\item{kde}{a list containing `x`- and `y`- arrays, and a `z`-matrix, as from MASS::kde2d (see below)}

\item{levels}{the required probability content of the contours (see below)}
}
\value{
the values of `z` that define the minimum-area regions with the specified probability contents
Given a list `kde` containing `x` and `y` (the x- and y-coordinates of grid points) and `z`
(values proportional to a probability density at those grid points), we calculate the
minimum-area contours which contain the probability contents specified in `levels`. All
the values in `levels` must be between 0 and 1.
}
\description{
Find values at which given KDE takes specified values
}
