% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{getPid}
\alias{getPid}
\title{Get the process id from a canonical cosmosis-timing filename. The
filename is expected to be in the form:
   cosmosis-timing-<pid>.db}
\usage{
getPid(fname)
}
\description{
Get the process id from a canonical cosmosis-timing filename. The
filename is expected to be in the form:
   cosmosis-timing-<pid>.db
}

