% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{load.cosmosis.timing.data}
\alias{load.cosmosis.timing.data}
\title{Load data from all the files in fnames. These are expected to be
cosmosis timing data files, with the canonical naming pattern:
   cosmosis-timing-<pid>.db}
\usage{
load.cosmosis.timing.data(fnames)
}
\description{
Returns a data.frame.
}

