% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.1d.likelihood.plots}
\alias{make.1d.likelihood.plots}
\title{Make a 1-d posterior density plot for each variable in the dataframe.
We use bw="nrd", which gives a bandwidth calculation according to:
   Scott, D. W. (1992) Multivariate Density Estimation: Theory, Practice, and
   Visualization. Wiley.}
\usage{
make.1d.likelihood.plots(dframe, prefix, output, device, verbose)
}
\description{
Make a 1-d posterior density plot for each variable in the dataframe.
We use bw="nrd", which gives a bandwidth calculation according to:
   Scott, D. W. (1992) Multivariate Density Estimation: Theory, Practice, and
   Visualization. Wiley.
}

