% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.2d.density.plot}
\alias{make.2d.density.plot}
\title{Make a 2-d density plot of  xcol vs. ycol, using data from df.}
\usage{
make.2d.density.plot(df, xcol, ycol, prefix, output, device, use.color, nbins)

make.2d.density.plot(df, xcol, ycol, prefix, output, device, use.color, nbins)
}
\description{
Make a 2-d density plot of  xcol vs. ycol, using data from df.

Make a 2-d density plot of  xcol vs. ycol, using data from df.
}

