% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.2d.density.plots}
\alias{make.2d.density.plots}
\title{For each pair, generate the kde2d result matrix.
Determine the values of z at which the 68% and 95% contour lines lie.
Transform the matrix to a dataframe.
Make the contour plot.}
\usage{
make.2d.density.plots(df, prefix, output, device, verbose, use.color, nbins)

make.2d.density.plots(df, prefix, output, device, verbose, use.color, nbins)
}
\description{
For each pair, generate the kde2d result matrix.
Determine the values of z at which the 68% and 95% contour lines lie.
Transform the matrix to a dataframe.
Make the contour plot.

For each pair, generate the kde2d result matrix.
Determine the values of z at which the 68% and 95% contour lines lie.
Transform the matrix to a dataframe.
Make the contour plot.
}

