% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.cmb.grand.spectrum.plot}
\alias{make.cmb.grand.spectrum.plot}
\title{Create the CMB spectrum plot with all spectra shown together. Because
some of the spectra may have negative values, we take the absolute
value (so that our log-log plot doesn't complain).}
\usage{
make.cmb.grand.spectrum.plot(dframe, prefix, outdir, verbose, devname, ...)
}
\description{
Create the CMB spectrum plot with all spectra shown together. Because
some of the spectra may have negative values, we take the absolute
value (so that our log-log plot doesn't complain).
}

