% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.cmb.spectrum.plot}
\alias{make.cmb.spectrum.plot}
\title{Create a single "CMB spectrum plot". A CMB spectrum plot is always a
plot of something as a function of C_ell. The plot is written to
a file, using the specified device. All additional arguments are
passed to the device function.}
\usage{
make.cmb.spectrum.plot(colname, dframe, prefix, outdir, verbose, devname, ...)
}
\description{
Create a single "CMB spectrum plot". A CMB spectrum plot is always a
plot of something as a function of C_ell. The plot is written to
a file, using the specified device. All additional arguments are
passed to the device function.
}

