% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.cmb.spectrum.plots}
\alias{make.cmb.spectrum.plots}
\title{Make all the CMB spectrum plots for files under the given top-level
directory *topdir*. CMB spectrum plots are made for all data files in
the subdirectory "cmb_cl".}
\usage{
make.cmb.spectrum.plots(topdir, verbose, prefix, outdir, devname)
}
\description{
Make all the CMB spectrum plots for files under the given top-level
directory *topdir*. CMB spectrum plots are made for all data files in
the subdirectory "cmb_cl".
}

