% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_input.R
\name{make.data.frame}
\alias{make.data.frame}
\title{Create a dataframe from CosmoSIS MCMC output. We expect the first line of
the output to contain the names of the parameters, separated by
spaces, and with section names separated from parameter names by a
double-hyphen.}
\usage{
make.data.frame(fname)
}
\arguments{
\item{fname}{The name of the file to be read.}

\item{burn}{The length of the burn-in period; these samples are ignored in making the dataframe.}
}
\value{
a CosmoSIS MCMC dataframe. The columns are:
  `l`
}
\description{
Create a dataframe from CosmoSIS MCMC output. We expect the first line of
the output to contain the names of the parameters, separated by
spaces, and with section names separated from parameter names by a
double-hyphen.
}

