% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.distance.plot}
\alias{make.distance.plot}
\title{Create a single "distance plot". A distance plot is always a plot of
something as a function of reshift "z'. The plot is written to a file,
using the specified device. All additional arguments are passed to the
device function.}
\usage{
make.distance.plot(colname, dframe, prefix, outdir, verbose, devname, ...)
}
\arguments{
\item{colname}{As a string, the name of the column from the dataframe to be used on the y-axis.}

\item{dframe}{A}
}
\value{
nothing; used for side-effect
}
\description{
Create a single "distance plot". A distance plot is always a plot of
something as a function of reshift "z'. The plot is written to a file,
using the specified device. All additional arguments are passed to the
device function.
}

