% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.distance.plots}
\alias{make.distance.plots}
\title{Make all the distance plots for files under the given top-level
directory *topdir*. Distance plots are made for data files in the
subdirectory "distances".}
\usage{
make.distance.plots(topdir, verbose, prefix, outdir, devname)
}
\description{
Make all the distance plots for files under the given top-level
directory *topdir*. Distance plots are made for data files in the
subdirectory "distances".
}

