% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{make.matterpower.dataframe}
\alias{make.matterpower.dataframe}
\title{Create a CosmoSIS matter power dataframe from all the named files.}
\usage{
make.matterpower.dataframe(dirname, type)
}
\arguments{
\item{dirname}{The name of the directory containing the CosmoSIS power
spectrum output.}

\item{type}{Any value; this value is replicated to fill the \code{type}
column of the dataframe.}
}
\value{
A CosmoSIS matter power dataframe. If the directory does not exist,
  the returned dataframe will be empty.
}
\description{
It is assumed the file format is the CosmoSIS format for storing power
spectra. Note this format is different from the format used for the storage
of scalar parameters.
}
