% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{make.theory.dataframe}
\alias{make.theory.dataframe}
\title{Create a CosmoSIS 'theory' dataframe from all the named files.}
\usage{
make.theory.dataframe(fnames)
}
\arguments{
\item{fnames}{A character vector containing the names of the files to be read.}
}
\value{
A CosmoSIS 'theory' dataframe.
}
\description{
Create a CosmoSIS 'theory' dataframe from the 'theory' files
generated by a run of CosmoSIS. It is assumed that each file contains
a single column of data, and that all the columns are of the same
length. This is the standard CosmoSIS output format for a variety of
theory calculations. Lines beginning with CosmoSIS comment character
"#" are ignored. The names of the columns are determined from the
names of the files, by dropping the file extension.
}
