% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{mcmc.list.from.emcee}
\alias{mcmc.list.from.emcee}
\title{mcmc.list.from.emcee}
\usage{
mcmc.list.from.emcee(tbl)
}
\arguments{
\item{tbl}{An emcee data.frame, as created by read.emcee}
}
\value{
an mcmc.list object
}
\description{
mcmc.list.from.emcee
}
