% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{mcmc.list.from.metropolis.hastings}
\alias{mcmc.list.from.metropolis.hastings}
\title{mcmc.list.from.metropolis.hastings}
\usage{
mcmc.list.from.metropolis.hastings(tbl)
}
\arguments{
\item{tbl}{A MH data.frame, as created by read.metropolis.hastings}
}
\value{
an mcmc.list object
}
\description{
mcmc.list.from.metropolis.hastings
}
