% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non.sampling.columns}
\alias{non.sampling.columns}
\title{non.sampling.columns}
\usage{
non.sampling.columns()
}
\value{
character The names of the columns in CosmoSIS output that we do *not* want to return.
}
\description{
non.sampling.columns
}
