% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_input.R
\name{parse.cosmosis.parameters}
\alias{parse.cosmosis.parameters}
\title{Extract the parameter names from a CosmoSIS sampler output file.}
\usage{
parse.cosmosis.parameters(txt)
}
\arguments{
\item{txt}{The text to be parsed.}
}
\value{
A character vector containing the names of the parameters read from
  the file.
}
\description{
The format of the first line of CosmoSIS grid and MCMC sample output carries
the names of the parameters that were varied in that run of CosmoSIS. This
function parses that line and returns the names of the parameters.
}

