% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read.cosmosis.grid}
\alias{read.cosmosis.grid}
\title{Read a CosmoSIS grid sampler output file.}
\usage{
read.cosmosis.grid(fname)
}
\arguments{
\item{fname}{The name of the CosmoSIS MCMC sampler output file to be read.}
}
\value{
A list of \code{n+1} components, where \code{n} is the number of
  coordinate axes comprising the grid.

  \describe{ \item{\code{x}, \code{y}}{The \code{x} and \code{y} coordinates
  of the grid points, vectors of length \code{m} and \code{n}.}
  \item{\code{z}}{An \code{m} by \code{n} matrix of log-likelihoods.} }
}
\description{
Reads a file in CosmoSIS grid sampler output format and returns a list
describing the data.
}
