% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read.cosmosis.mcmc}
\alias{read.cosmosis.mcmc}
\title{Create a data frame from CosmoSIS MCMC sampler output.}
\usage{
read.cosmosis.mcmc(fname, burn = 0)
}
\arguments{
\item{fname}{The name of the CosmoSIS MCMC sampler output file to be read.
making the data frame}

\item{burn}{The length of the burn-in period; these samples are ignored in
making the data frame.}
}
\value{
a CosmoSIS MCMC data frame
}
\description{
Reads a file in CosmoSIS MCMC sampler output format and creates a data frame
from it. Each line in the file corresponds to a sample in the data frame;
each column in the file corresponds to a column in the data frame.
}
\details{
The columns in a CosmoSIS MCMC data frame are:
\describe{
\item{loglike}{log-likelihood of the sample.}
 \item{like}{normalized likelihood of the sample.}
 \item{\emph{others}}{one column per sampled variable in the MCMC
  output, named as in the output.}
}

We expect the first line of the output to contain the names of the
parameters, separated by spaces, and with section names separated from
parameter names by a double-hyphen.
}
