% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read.emcee}
\alias{read.emcee}
\title{read.emcee}
\usage{
read.emcee(fname)
}
\arguments{
\item{fname}{The name of the CosmoSIS MCMC sampler output file to be read.
making the data frame}
}
\value{
a CosmoSIS MCMC dataframe
}
\description{
read.emcee
}
