% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read.metropolis.hastings}
\alias{read.metropolis.hastings}
\title{read.metropolis.hastings}
\usage{
read.metropolis.hastings(fileglob)
}
\arguments{
\item{fileglob}{The glob pattern (as used by Sys.glob) specifiying the CosmoSIS
Metropolis-Hastings sampler output file to be read.}
}
\value{
a CosmoSIS MCMC dataframe
}
\description{
read.metropolis.hastings
}
