% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read.multinest}
\alias{read.multinest}
\title{read.multinest}
\usage{
read.multinest(fname, remove.small = TRUE)
}
\arguments{
\item{fname}{The name of the CosmoSIS output file to be read}

\item{remove.small}{If TRUE, remove very small weights}
}
\value{
a weighted CosmoSIS dataframe
}
\description{
read.multinest
}
