% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{remove.burnin}
\alias{remove.burnin}
\title{remove.burnin}
\usage{
remove.burnin(x, n)
}
\arguments{
\item{x}{: a emcee or MH dataframe}

\item{n}{: the number of samples to remove (from each walker or chain)}
}
\value{
a copy of the input x, with n samples removed
}
\description{
remove.burnin
}
