% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_basics.R
\name{scales.log.log}
\alias{scales.log.log}
\title{Convenience functions for drawing log axes with non-default tick
positions and labels.}
\usage{
scales.log.log(xbase = 10, ybase = xbase, xeq = FALSE, yeq = xeq)
}
\arguments{
\item{xbase}{The base to be used for the x-axis.}

\item{ybase}{The base to be used for the y-axis.}

\item{xeq}{\code{equispaced} option for x-axis (TRUE or FALSE).}

\item{yeq}{\code{equispaced} option for y-axis (TRUE or FALSE).}
}
\value{
A list suitable for the \code{scales} argument of Lattice
plotting functions, e.g. \code{xyplot}.
}
\description{
Convenience functions for drawing log axes with non-default tick
positions and labels.
}

