% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{vmat2df}
\alias{vmat2df}
\title{vmat2df will convert the kind of list returned by kde2d (containing two)
vectors and a matrix, named x, y and z) into a dataframe with columns
x, y, z.}
\usage{
vmat2df(u)

vmat2df(u)
}
\description{
vmat2df will convert the kind of list returned by kde2d (containing two)
vectors and a matrix, named x, y and z) into a dataframe with columns
x, y, z.

vmat2df will convert the kind of list returned by kde2d (containing two)
vectors and a matrix, named x, y and z) into a dataframe with columns
x, y, z.
}

