% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_basics.R
\name{xscale.log}
\alias{xscale.log}
\title{Return the a function that returns the function
\code{xscale.components.log10ticks}, for more convient use in
\code{xyplot}.}
\usage{
xscale.log()
}
\description{
Return the a function that returns the function
\code{xscale.components.log10ticks}, for more convient use in
\code{xyplot}.
}

