% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_basics.R
\name{yscale.log}
\alias{yscale.log}
\title{Return the a function that returns the function
\code{yscale.components.log10ticks}, for more convient use in
\code{xyplot}.}
\usage{
yscale.log()
}
\description{
Return the a function that returns the function
\code{yscale.components.log10ticks}, for more convient use in
\code{xyplot}.
}

