% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append.likelihoods}
\alias{append.likelihoods}
\title{Append (normalized) likelihoods to a data.frame containing log-likelihoods.}
\usage{
append.likelihoods(d)
}
\arguments{
\item{d}{A data.frame containing a column of log-likelihoods named loglike or like.}
}
\value{
The augmented data.frame, with a column `like` containing likelihoods
}
\description{
Append (normalized) likelihoods to a data.frame containing log-likelihoods.
}
