% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols2vmat}
\alias{cols2vmat}
\title{Transform a dataframe carrying grid-oriented data into a list of grid
vectors and a data matrix.}
\usage{
cols2vmat(d)
}
\arguments{
\item{d}{A data.frame with 3 columns. The first two columns must be vectors of grid
coordinates, and the final column contains the value at that grid coordinate.}
}
\value{
a list of length 3; the first 2 entries are vectors of the grid values in each
  direction, and the last entry is the 2 dimensional matrix of the data
  values.
}
\description{
Transform a dataframe carrying grid-oriented data into a list of grid
vectors and a data matrix.
}
