% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{find.contours}
\alias{find.contours}
\title{Return the likelihood values corresponding to the boundary of the
regions containing the probability contents 'levels'.}
\usage{
find.contours(kde, levels = c(0.68, 0.95))
}
\description{
Return the likelihood values corresponding to the boundary of the
regions containing the probability contents 'levels'.
}

