% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{make.data.frame}
\alias{make.data.frame}
\title{Create a dataframe from CosmoSIS output. We expect the first line of
the output to contain the names of the parameters, separated by
spaces, and with section names separated from parameter names by a
double-hyphen.}
\usage{
make.data.frame(fname, burn)
}
\description{
Create a dataframe from CosmoSIS output. We expect the first line of
the output to contain the names of the parameters, separated by
spaces, and with section names separated from parameter names by a
double-hyphen.
}

