% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_density_2d}
\alias{plot_density_2d}
\title{plot_density_2d}
\usage{
plot_density_2d(
  data,
  x,
  y,
  bins = 50,
  scale_x = scale_x_continuous,
  scale_y = scale_y_continuous,
  levels = c(0.6826895, 0.9544997, 0.9973002)
)
}
\arguments{
\item{data}{a CosmoSIS MCMC dataframe}

\item{x}{the variable to be plotted on the x-axis}

\item{y}{the variable to be plotted on the y-axis}

\item{bins}{the numbers of bins for both `x` and `y`}

\item{scale_x}{position scale function for `x`}

\item{scale_y}{position scale function for `y`}

\item{levels}{probability content to be included in each contour line}
}
\value{
a ggplot object
}
\description{
plot_density_2d
}
\examples{
\dontrun{
  dplyr::filter(samples, sample > nburn) \%>\% plot_density_2d(omega_m, sigma8_input)
}
}
