% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_density_2d}
\alias{plot_density_2d}
\title{Produce a 2-d posterior density plot}
\usage{
plot_density_2d(
  data,
  x,
  y,
  bins = 50,
  levels = c(0.6826895, 0.9544997, 0.9973002)
)
}
\arguments{
\item{data}{a CosmoSIS MCMC dataframe}

\item{x}{the variable to be plotted on the x-axis}

\item{y}{the variable to be plotted on the y-axis}

\item{bins}{the numbers of bins for both `x` and `y`}

\item{levels}{probability content to be included in each contour line}
}
\value{
a ggplot object
}
\description{
Produce a 2-d posterior density plot, using a combined hexbin and contour plot,
including gradient coloring to show the relative density.
}
\examples{
\dontrun{
  dirname <- system.file("extdata", "run20", package = "rcosmosis")
  fglob <- file.path(dirname, "chain_metro_20_*.txt.xz")
  samples <- read.metropolis.hastings(fglob)
  dplyr::filter(samples, sample > 100) \%>\% plot_density_2d(omega_m, sigma8_input)
}
}
