% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcosmosis-package.R
\docType{package}
\name{rcosmosis}
\alias{rcosmosis}
\alias{rcosmosis-package}
\title{rcosmosis: A package for analysis of CosmoSIS output.}
\description{
The rcosmosis package provides functions for reading CosmoSIS output (MCMC
sampler and grid output files), and for making 1- and 2-dimensional plots of
that output.
}
\section{Functions for reading data}{

\code
}

\section{Functions for plotting data}{

 \emph{Yet to be written.}
}

