% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vmat2df}
\alias{vmat2df}
\title{convert output of kde2d into a dataframe}
\usage{
vmat2df(u)
}
\arguments{
\item{u}{the kde2d result to be converted}
}
\value{
a dataframe with columns `x`, `y` and `z`.
}
\description{
convert output of kde2d into a dataframe
}
